requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
astroid==2.15.0
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
commonmark==0.9.1
dill==0.3.6
flake8==6.0.0
frozenlist==1.3.3
gTTS==2.3.1
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
multidict==6.0.4
openai==0.27.0
platformdirs==3.1.1
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.14.0
pylint==2.17.0
python-dotenv==1.0.0
requests==2.28.2
rich==12.6.0
shellingham==1.5.0.post1
tomlkit==0.11.6
tqdm==4.65.0
typer==0.7.0
urllib3==1.26.14
wrapt==1.15.0
yarl==1.8.2
